pyproject.toml
[tool.poetry]
name = "kotanima_server"
version = "0.1.0"
description = ""
authors = ["b0und <f1arenova1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.3.1"
cached-property = "1.5.2"
certifi = "2020.12.5"
cffi = "1.14.5"
chardet = "4.0.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "3.4.4"
defusedxml = "0.7.0rc2"
Django = "3.1.6"
django-templated-mail = "1.1.1"
django-url-filter = "0.3.15"
djangorestframework = "3.12.2"
djangorestframework-simplejwt = "4.6.0"
djoser = "2.1.0"
enum-compat = "0.0.3"
idna = "2.10"
itypes = "1.2.0"
Jinja2 = "2.11.3"
MarkupSafe = "1.1.1"
oauthlib = "3.1.0"
psycopg2-binary = "2.8.6"
pycparser = "2.20"
PyJWT = "2.0.1"
python3-openid = "3.2.0"
pytz = "2021.1"
requests = "2.25.1"
requests-oauthlib = "1.3.0"
six = "1.15.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.0.3"
sqlparse = "0.4.1"
uritemplate = "3.0.1"
urllib3 = "1.26.3"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
